Modularize simulations
