+++ openfisca_core/errors/__init__.py
@@ -21,9 +21,12 @@
 #
 # See: https://www.python.org/dev/peps/pep-0008/#imports
 
+from .cycle_error import CycleError  # noqa: F401
 from .empty_argument_error import EmptyArgumentError  # noqa: F401
+from .nan_creation_error import NaNCreationError  # noqa: F401
 from .parameter_not_found_error import ParameterNotFoundError, ParameterNotFoundError as ParameterNotFound  # noqa: F401
 from .parameter_parsing_error import ParameterParsingError  # noqa: F401
 from .period_mismatch_error import PeriodMismatchError  # noqa: F401
 from .situation_parsing_error import SituationParsingError  # noqa: F401
+from .spiral_error import SpiralError  # noqa: F401
 from .variable_not_found_error import VariableNotFoundError, VariableNotFoundError as VariableNotFound  # noqa: F401
