+++ openfisca_core/errors/cycle_error.py
@@ -1,0 +1,4 @@
+class CycleError(Exception):
+    """Simulation error."""
+
+    pass
