+++ openfisca_core/errors/nan_creation_error.py
@@ -1,0 +1,4 @@
+class NaNCreationError(Exception):
+    """Simulation error."""
+
+    pass
