+++ openfisca_core/errors/spiral_error.py
@@ -1,0 +1,4 @@
+class SpiralError(Exception):
+    """Simulation error."""
+
+    pass
