+++ openfisca_core/simulations/helpers.py
@@ -1,0 +1,6 @@
+def calculate_output_add(simulation, variable_name, period):
+    return simulation.calculate_add(variable_name, period)
+
+
+def calculate_output_divide(simulation, variable_name, period):
+    return simulation.calculate_divide(variable_name, period)
